refactor: 커뮤니티 구조 리팩토링
