+++ src/stores/useTabStore.tsx
@@ -1,0 +1,11 @@
+import { create } from 'zustand';
+
+interface TabState {
+  activeTab: string;
+  setActiveTab: (tab: string) => void;
+}
+
+export const useTabStore = create<TabState>((set) => ({
+  activeTab: 'QnA',
+  setActiveTab: (tab) => set({ activeTab: tab }),
+}));
